# docs/contracts/runtime-api-contract.v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "投资进化系统 Runtime Contract",
    "version": "1.1.0",
    "description": "Derived OpenAPI document generated from docs/contracts/runtime-api-contract.v1.json"
  },
  "servers": [
    {
      "url": "/",
      "description": "Project root; paths remain absolute and unversioned."
    }
  ],
  "paths": {
...
    "/api/chat": {
      "post": {
        "operationId": "chat_ask",
        "tags": [
          "chat"
        ],
        "summary": "Send a natural-language message to Commander.",
        "description": "Provide session_key/chat_id to continue a multi-turn API conversation; omit them for isolated one-shot requests.",
        "responses": {
          "200": {
            "description": "Send a natural-language message to Commander.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/chatReply"
                }
              }
            }
          },
          "400": {
            "description": "message is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "503": {
            "description": "Commander runtime is not initialized or background loop is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected command/runtime error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          }
        },
        "x-runtime-required": true,
        "x-runtime-preferred": true,
        "x-latency": "medium",
        "x-realtime": false,
        "x-pagination": "none",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string",
                    "minLength": 1
                  },
                  "session_key": {
                    "type": "string",
                    "minLength": 1
                  },
                  "chat_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "request_id": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "operationId": "chat_stream",
        "tags": [
          "chat"
        ],
        "summary": "Open a session-bound SSE stream for runtime events plus the final Commander reply.",
        "description": "The stream first emits a connected event, then request-bound runtime events, then the final reply and done event.\nThis is event streaming for runtime state; it does not stream LLM tokens.",
        "responses": {
          "200": {
            "description": "Open a session-bound SSE stream for runtime events plus the final Commander reply.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "message is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "503": {
            "description": "Commander runtime is not initialized or background loop is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected command/runtime error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          }
        },
        "x-runtime-required": true,
        "x-runtime-preferred": true,
        "x-latency": "medium",
        "x-realtime": true,
        "x-pagination": "none",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string",
                    "minLength": 1
                  },
                  "session_key": {
                    "type": "string",
                    "minLength": 1
                  },
                  "chat_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "request_id": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/train": {
      "post": {
        "operationId": "training_direct",
        "tags": [
          "training"
        ],
        "summary": "Convenience shortcut: auto-create a plan and execute it immediately.",
        "description": "mock=false by default; mock is an explicit smoke/demo mode and no longer an implicit fallback.",
        "responses": {
          "200": {
            "description": "Convenience shortcut: auto-create a plan and execute it immediately.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/trainingExecution"
                }
              }
            }
          },
          "400": {
            "description": "rounds must be an integer; mock must be a boolean",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "503": {
            "description": "Commander runtime is not initialized or background loop is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected training execution error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          }
        },
        "x-runtime-required": true,
        "x-runtime-preferred": false,
        "x-latency": "long",
        "x-realtime": true,
        "x-pagination": "none",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rounds": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "maximum": 100
                  },
                  "mock": {
                    "type": "boolean",
                    "default": false,
                    "accepts_bool_like_strings": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/lab/training/plans": {
      "post": {
        "operationId": "trainingPlans_create",
        "tags": [
          "training_lab"
        ],
        "summary": "Create a structured training plan artifact without executing it.",
        "description": "Create a structured training plan artifact without executing it.",
        "responses": {
          "201": {
            "description": "Create a structured training plan artifact without executing it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/trainingPlan"
                }
              }
            }
          },
          "400": {
            "description": "rounds must be an integer; mock must be a boolean; detail_mode must be one of: fast, slow; tags must be a list of strings or a comma-separated string",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "503": {
            "description": "Commander runtime is not initialized or background loop is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          }
        },
        "x-runtime-required": true,
        "x-runtime-preferred": true,
        "x-latency": "fast",
        "x-realtime": false,
        "x-pagination": "none",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rounds": {
                    "type": "integer",
                    "default": 1,
                    "minimum": 1,
                    "maximum": 100
                  },
                  "mock": {
                    "type": "boolean",
                    "default": false,
                    "accepts_bool_like_strings": true
                  },
                  "goal": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  },
                  "tags": {
                    "type": [
                      "array",
                      "string"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "detail_mode": {
                    "type": "string",
                    "default": "fast",
                    "enum": [
                      "fast",
                      "slow"
                    ]
                  },
                  "protocol": {
                    "type": "object"
                  },
                  "dataset": {
                    "type": "object"
                  },
                  "model_scope": {
                    "type": "object"
                  },
                  "optimization": {
                    "type": "object"
                  },
                  "llm": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "trainingPlans_list",
        "tags": [
          "training_lab"
        ],
        "summary": "List latest training plan artifacts.",
        "description": "List latest training plan artifacts.",
        "responses": {
          "200": {
            "description": "List latest training plan artifacts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/artifactList"
                }
              }
            }
          },
          "400": {
            "description": "limit must be an integer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "503": {
            "description": "Commander runtime is not initialized or background loop is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          }
        },
        "x-runtime-required": true,
        "x-runtime-preferred": true,
        "x-latency": "fast",
        "x-realtime": false,
        "x-pagination": "limit-only",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": ""
          }
        ]
      }
    },
...
    "/api/lab/training/plans/{plan_id}/execute": {
      "post": {
        "operationId": "trainingPlans_execute",
        "tags": [
          "training_lab"
        ],
        "summary": "Execute a persisted training plan.",
        "description": "Execute a persisted training plan.",
        "responses": {
          "200": {
            "description": "Execute a persisted training plan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/trainingExecution"
                }
              }
            }
          },
          "503": {
            "description": "Commander runtime is not initialized or background loop is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "404": {
            "description": "training plan not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected training plan execution error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flatError"
                }
              }
            }
          }
        },
        "x-runtime-required": true,
        "x-runtime-preferred": true,
        "x-latency": "long",
        "x-realtime": true,
        "x-pagination": "none",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "flatError": {
        "type": "object",
        "description": "Most current endpoints return a flat error body on non-2xx.",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "chatReply": {
        "type": "object",
        "properties": {
          "reply": {
            "type": "string"
          },
          "next_action": {
            "$ref": "#/components/schemas/responseNextAction"
          },
          "feedback": {
            "$ref": "#/components/schemas/responseFeedback"
          },
          "message": {
            "type": "string"
          },
          "task_bus": {
            "type": "object",
            "additionalProperties": true
          },
          "session_key": {
            "type": "string"
          },
          "chat_id": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "reply"
        ],
        "additionalProperties": false
      },
      "trainingPlan": {
        "type": "object",
        "properties": {
          "plan_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "auto_generated": {
            "type": "boolean"
          },
          "spec": {
            "type": "object",
            "additionalProperties": true
          },
          "protocol": {
            "type": "object",
            "additionalProperties": true
          },
          "dataset": {
            "type": "object",
            "additionalProperties": true
          },
          "model_scope": {
            "type": "object",
            "additionalProperties": true
          },
          "optimization": {
            "type": "object",
            "additionalProperties": true
          },
          "llm": {
            "type": "object",
            "additionalProperties": true
          },
          "objective": {
            "type": "object",
            "additionalProperties": true
          },
          "artifacts": {
            "type": "object",
            "additionalProperties": true
          },
          "last_run_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_run_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "feedback": {
            "$ref": "#/components/schemas/responseFeedback"
          },
          "next_action": {
            "$ref": "#/components/schemas/responseNextAction"
          }
        },
        "required": [
          "plan_id",
          "created_at",
          "status",
          "source",
          "spec",
          "protocol",
          "dataset",
          "model_scope",
          "optimization",
          "llm",
          "objective",
          "artifacts"
        ],
        "additionalProperties": true
      },
      "artifactList": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": [
          "count",
          "items"
        ],
        "additionalProperties": false
      },
...
      "trainingExecution": {
        "type": "object",
        "description": "Result of executing a training plan or the convenience /api/train shortcut.",
        "properties": {
          "status": {
            "type": "string"
          },
          "rounds": {
            "type": "integer"
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "summary": {
            "type": "object",
            "additionalProperties": true
          },
          "training_lab": {
            "type": "object",
            "additionalProperties": true
          },
          "feedback": {
            "$ref": "#/components/schemas/responseFeedback"
          },
          "next_action": {
            "$ref": "#/components/schemas/responseNextAction"
          }
        },
        "required": [
          "status",
          "results",
          "summary",
          "training_lab"
        ],
        "additionalProperties": true
      },
...
      "responseFeedback": {
        "type": "object",
        "description": "Standard cross-domain feedback fragment returned by Commander-backed workflows and stock-analysis responses.",
        "properties": {
          "message": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "reason_codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason_texts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "requires_confirmation": {
            "type": "boolean"
          },
          "decision": {
            "type": "string"
          },
          "coverage": {
            "type": "object",
            "properties": {
              "planned_step_coverage": {
                "type": "number"
              },
              "parameter_coverage": {
                "type": "number"
              }
            },
            "required": [
              "planned_step_coverage",
              "parameter_coverage"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "message",
          "summary",
          "reason_codes",
          "reason_texts",
          "requires_confirmation",
          "decision",
          "coverage"
        ],
        "additionalProperties": true
      },
      "responseNextAction": {
        "type": "object",
        "description": "Standard cross-domain next-action fragment derived from feedback and gate state.",
        "properties": {
          "kind": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "requires_confirmation": {
            "type": "boolean"
          },
          "suggested_params": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "kind",
          "label",
          "description",
          "requires_confirmation",
          "suggested_params"
        ],
        "additionalProperties": true
      },
...
    }
  },
...
}